Move hosted -> IoP re-registration test to destructive module
After merging https://github.com/SatelliteQE/robottelo/pull/21054, I
realized that, while it passes when run by itself, the test added by
that PR will encounter collisions when run with the whole
`ui/test_rhcloud_iop.py` module. All of the other tests in that module
require starting with an IoP-enabled Sat, while the test from #21054
requires starting with a Sat configured for hosted Insights and only
enabling IoP after registering a host to hosted Insights. This PR
isolates the test to a module in the `destructive` directory to avoid
any collisions.
